fixed all warnings on /W3... compiles cleanly on warning level 3. However, warning C4715 is suppressed because boost. Not all control paths return a value is the warning, however none of our code has that warning.

+++ Dev/StaticSheepNeoBuildSystem/game/SheepEngine/Components/controllers/player/CombatController.cpp
@@ -89,7 +89,7 @@ namespace Framework
     if (bullet)
     {
       /* Inflict the damage to the player (or atleast attempt) */
-      TakeDamage(bullet->damage, bullet->playerOwner);
+      TakeDamage(static_cast<float>(bullet->damage), bullet->playerOwner);
 
       /* Play hit sound */
       se->Play("hit1", &SoundInstance(1.0f));
